# docs/openapi/catalog-api-v1.json
{
  "components": {
    "parameters": {
      "RequestId": {
        "description": "Optional client correlation ID. The service generates and returns one when omitted.",
        "in": "header",
        "name": "X-Request-ID",
        "required": false,
        "schema": {
          "maxLength": 128,
          "type": "string"
        }
      }
    },
    "schemas": {
...
      "AssetRead": {
        "properties": {
          "asset_type": {
            "$ref": "#/components/schemas/AssetType"
          },
          "classification": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Classification"
          },
          "columns": {
            "items": {
              "$ref": "#/components/schemas/AssetColumnRead"
            },
            "title": "Columns",
            "type": "array"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "curated_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Curated At"
          },
          "database_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Database Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "discovery_score": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Discovery Score"
          },
          "domain_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Domain Id"
          },
          "freshness_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Freshness At"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "last_discovered_at": {
            "format": "date-time",
            "title": "Last Discovered At",
            "type": "string"
          },
          "lifecycle_status": {
            "$ref": "#/components/schemas/LifecycleStatus"
          },
          "metadata_versions": {
            "items": {
              "$ref": "#/components/schemas/AssetMetadataVersionRead"
            },
            "title": "Metadata Versions",
            "type": "array"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "owner": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner"
          },
          "qualified_name": {
            "title": "Qualified Name",
            "type": "string"
          },
          "quality_explainable_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quality Explainable At"
          },
          "quality_score": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quality Score"
          },
          "row_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Row Count"
          },
          "schema_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Schema Name"
          },
          "source_id": {
            "title": "Source Id",
            "type": "string"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "title": "Tags",
            "type": "array"
          },
          "technical_description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Technical Description"
          },
          "technical_metadata": {
            "additionalProperties": true,
            "title": "Technical Metadata",
            "type": "object"
          },
          "technical_version": {
            "title": "Technical Version",
            "type": "integer"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "source_id",
          "asset_type",
          "qualified_name",
          "name",
          "database_name",
          "schema_name",
          "technical_description",
          "technical_metadata",
          "row_count",
          "last_discovered_at",
          "freshness_at",
          "technical_version",
          "description",
          "tags",
          "owner",
          "classification",
          "domain_id",
          "quality_score",
          "quality_explainable_at",
          "lifecycle_status",
          "curated_at",
          "created_at",
          "updated_at"
        ],
        "title": "AssetRead",
        "type": "object"
      },
      "AssetSearchResponse": {
        "properties": {
          "facets": {
            "additionalProperties": {
              "additionalProperties": {
                "type": "integer"
              },
              "type": "object"
            },
            "title": "Facets",
            "type": "object"
          },
          "index_fresh_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Index Fresh At"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/AssetRead"
            },
            "title": "Items",
            "type": "array"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total"
        ],
        "title": "AssetSearchResponse",
        "type": "object"
      },
...
      "ErrorDetail": {
        "properties": {
          "code": {
            "description": "Stable machine-readable application error code.",
            "type": "string"
          },
          "details": {
            "description": "Optional validation or domain-specific evidence."
          },
          "message": {
            "description": "Safe human-readable error message.",
            "type": "string"
          },
          "request_id": {
            "description": "Correlation identifier for support and audit investigation.",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "request_id"
        ],
        "type": "object"
      },
      "ErrorEnvelope": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "required": [
          "error"
        ],
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "bearerFormat": "JWT",
        "description": "JWT or OIDC access token with a tenant claim and one or more DataGenie roles.",
        "scheme": "bearer",
        "type": "http"
      },
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {
      "name": "DataGenie API Support",
      "url": "https://github.com/itismohan/datagenie/issues"
    },
    "description": "## DataGenie Catalog API\n\nThe DataGenie Catalog API provides tenant-isolated metadata discovery, stewardship,\ngovernance, operational lineage integration boundaries, quality-aware search, and\ncustomer operations. Every endpoint is versioned under `/api/v1`.\n\n### Authentication and tenant boundary\n\nSupply an OAuth/OIDC or HS256 bearer token in `Authorization: Bearer <token>`.\nOutside local development, every token must include the configured tenant claim.\nResponses, search facets, audit history, exports, retention policies, and webhook\noperations are restricted to that tenant.\n\n### Request correlation and errors\n\nClients may send `X-Request-ID`; otherwise the service generates one and returns it\non every response. Error responses use a stable envelope containing an application\n`code`, a human-readable `message`, and the request ID required for support.\n\n### Idempotency and rate limits\n\nUse `Idempotency-Key` for supported mutating operations, especially source and\ningestion-job creation. Catalog endpoints return standard `RateLimit-*` and\n`Retry-After` headers when request protection is active.\n",
    "license": {
      "name": "Proprietary"
    },
    "summary": "Tenant-isolated metadata governance and discovery platform",
    "title": "DataGenie Catalog API",
    "version": "1.0.0",
...
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/assets/": {
      "get": {
        "operationId": "list_assets_api_v1_assets__get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 255,
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "source_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Source Id"
            }
          },
          {
            "in": "query",
            "name": "asset_type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Asset Type"
            }
          },
          {
            "in": "query",
            "name": "lifecycle_status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Lifecycle Status"
            }
          },
          {
            "in": "query",
            "name": "owner",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Owner"
            }
          },
          {
            "in": "query",
            "name": "classification",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Classification"
            }
          },
          {
            "in": "query",
            "name": "tag",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tag"
            }
          },
          {
            "in": "query",
            "name": "freshness_before",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Freshness Before"
            }
          },
          {
            "in": "query",
            "name": "domain",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Domain"
            }
          },
          {
            "in": "query",
            "name": "business_term",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Business Term"
            }
          },
          {
            "in": "query",
            "name": "quality_min",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 100,
                  "minimum": 0,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Quality Min"
            }
          },
          {
            "in": "query",
            "name": "explainable_quality_only",
            "required": false,
            "schema": {
              "default": false,
              "title": "Explainable Quality Only",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "discovery_session_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 128,
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Discovery Session Id"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "$ref": "#/components/parameters/RequestId"
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssetSearchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "examples": {
                  "unauthorized": {
                    "value": {
                      "error": {
                        "code": "unauthorized",
                        "message": "A valid bearer token is required.",
                        "request_id": "b9ce11cc-4e68-4f4c-9e0f-4c1f0e2c6e2d"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "A valid bearer token is required."
          },
          "403": {
            "content": {
              "application/json": {
                "examples": {
                  "forbidden": {
                    "value": {
                      "error": {
                        "code": "forbidden",
                        "message": "The authenticated principal is not authorized for this operation.",
                        "request_id": "b9ce11cc-4e68-4f4c-9e0f-4c1f0e2c6e2d"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The authenticated principal is not authorized for this operation."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "429": {
            "content": {
              "application/json": {
                "examples": {
                  "rate_limit_exceeded": {
                    "value": {
                      "error": {
                        "code": "rate_limit_exceeded",
                        "message": "Too many requests. Retry after the supplied interval.",
                        "request_id": "b9ce11cc-4e68-4f4c-9e0f-4c1f0e2c6e2d"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Too many requests. Retry after the supplied interval."
          },
          "500": {
            "content": {
              "application/json": {
                "examples": {
                  "internal_error": {
                    "value": {
                      "error": {
                        "code": "internal_error",
                        "message": "An unexpected server error occurred.",
                        "request_id": "b9ce11cc-4e68-4f4c-9e0f-4c1f0e2c6e2d"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "An unexpected server error occurred."
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "List Assets",
        "tags": [
          "Assets"
        ]
      }
    },
    "/api/v1/assets/{asset_id}": {
      "get": {
        "operationId": "get_asset_api_v1_assets__asset_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "asset_id",
            "required": true,
            "schema": {
              "title": "Asset Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "discovery_session_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 128,
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Discovery Session Id"
            }
          },
          {
            "in": "query",
            "name": "purpose",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 500,
                  "minLength": 3,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Purpose"
            }
          },
          {
            "$ref": "#/components/parameters/RequestId"
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssetRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "examples": {
                  "unauthorized": {
                    "value": {
                      "error": {
                        "code": "unauthorized",
                        "message": "A valid bearer token is required.",
                        "request_id": "b9ce11cc-4e68-4f4c-9e0f-4c1f0e2c6e2d"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "A valid bearer token is required."
          },
          "403": {
            "content": {
              "application/json": {
                "examples": {
                  "forbidden": {
                    "value": {
                      "error": {
                        "code": "forbidden",
                        "message": "The authenticated principal is not authorized for this operation.",
                        "request_id": "b9ce11cc-4e68-4f4c-9e0f-4c1f0e2c6e2d"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The authenticated principal is not authorized for this operation."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "429": {
            "content": {
              "application/json": {
                "examples": {
                  "rate_limit_exceeded": {
                    "value": {
                      "error": {
                        "code": "rate_limit_exceeded",
                        "message": "Too many requests. Retry after the supplied interval.",
                        "request_id": "b9ce11cc-4e68-4f4c-9e0f-4c1f0e2c6e2d"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Too many requests. Retry after the supplied interval."
          },
          "500": {
            "content": {
              "application/json": {
                "examples": {
                  "internal_error": {
                    "value": {
                      "error": {
                        "code": "internal_error",
                        "message": "An unexpected server error occurred.",
                        "request_id": "b9ce11cc-4e68-4f4c-9e0f-4c1f0e2c6e2d"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "An unexpected server error occurred."
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Get Asset",
        "tags": [
          "Assets"
        ]
      },
      "patch": {
        "operationId": "curate_asset_api_v1_assets__asset_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "asset_id",
            "required": true,
            "schema": {
              "title": "Asset Id",
              "type": "string"
            }
          },
          {
            "$ref": "#/components/parameters/RequestId"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssetCurationUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssetRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "examples": {
                  "unauthorized": {
                    "value": {
                      "error": {
                        "code": "unauthorized",
                        "message": "A valid bearer token is required.",
                        "request_id": "b9ce11cc-4e68-4f4c-9e0f-4c1f0e2c6e2d"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "A valid bearer token is required."
          },
          "403": {
            "content": {
              "application/json": {
                "examples": {
                  "forbidden": {
                    "value": {
                      "error": {
                        "code": "forbidden",
                        "message": "The authenticated principal is not authorized for this operation.",
                        "request_id": "b9ce11cc-4e68-4f4c-9e0f-4c1f0e2c6e2d"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The authenticated principal is not authorized for this operation."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "429": {
            "content": {
              "application/json": {
                "examples": {
                  "rate_limit_exceeded": {
                    "value": {
                      "error": {
                        "code": "rate_limit_exceeded",
                        "message": "Too many requests. Retry after the supplied interval.",
                        "request_id": "b9ce11cc-4e68-4f4c-9e0f-4c1f0e2c6e2d"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Too many requests. Retry after the supplied interval."
          },
          "500": {
            "content": {
              "application/json": {
                "examples": {
                  "internal_error": {
                    "value": {
                      "error": {
                        "code": "internal_error",
                        "message": "An unexpected server error occurred.",
                        "request_id": "b9ce11cc-4e68-4f4c-9e0f-4c1f0e2c6e2d"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "An unexpected server error occurred."
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Curate Asset",
        "tags": [
          "Assets"
        ]
      }
    },
...
  },
  "servers": [
...
}